refactored text into chunks? kind of
