+++ src/components/text-animation/text-animator.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect } from "react";
+import { shuffleArray } from "./shuffle-array";
 
-// props for tsx
-type TextAnimationProps = {
+// custom hook split for reuse & organization
+type UseTextAnimationProps = {
 	initialSentences: string[];
 	loopSentences: string[];
 	fadeTrue: boolean;
@@ -10,33 +11,20 @@ type TextAnimationProps = {
 	fadeDuration?: number;
 };
 
-// fisher yates shuffle since the previous one was too archaic
-const shuffleArray = (array: string[]): string[] => {
-	// input is an array of strings, output an array of strings
-	const shuffled = [...array]; // clone array
-	for (let i = shuffled.length - 1; i > 0; i--) {
-		// reverse loop from end to beginning
-		const j = Math.floor(Math.random() * (i + 1)); // random index between 0 and i
-		[shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]]; // destructive swap / swap i element with j element
-	}
-	return shuffled;
-};
-
-const TextAnimation = ({
+export const useTextAnimation = ({
 	initialSentences, // array of sentences to be inputted
 	loopSentences,
 	fadeTrue, // if true, should fade automatically, if not, will fade when step ends
 	speed = 60,
 	delayBetweenSentences = 1500,
 	fadeDuration = 2000,
-}: TextAnimationProps) => {
+}: UseTextAnimationProps) => {
 	const [displayedText, setDisplayedText] = useState<string>(""); // useState for the displayed text
 	const [currentSentenceIndex, setCurrentSentenceIndex] = useState<number>(0); // use state for the current index of the sentence (if multiple sentences used)
 	const [fadingOut, setFadingOut] = useState<boolean>(false); // fade out boolean
 	const [isInLoopPhase, setIsInLoopPhase] = useState<boolean>(false); // determine whether the initial sentence array has been run through
 	const [currentLoopSentences, setCurrentLoopSentences] =
-		useState(loopSentences);
-	// storing the new sentence array after being shuffled
+		useState(loopSentences); // storing the new sentence array after being shuffled
 
 	// flip flop between sentence array
 	const sentences = isInLoopPhase ? currentLoopSentences : initialSentences;
@@ -45,17 +33,14 @@ const TextAnimation = ({
 	useEffect(() => {
 		// if the index reaches end of array, just end/return
 		if (currentSentenceIndex >= sentences.length) {
-			console.log(loopSentences);
 			// transition to loop sentences after initial ones
 			if (!isInLoopPhase && loopSentences.length > 0) {
 				setCurrentLoopSentences(shuffleArray(loopSentences));
 				setCurrentSentenceIndex(0);
 				setIsInLoopPhase(true);
-				console.log("shuffled", currentLoopSentences);
 			} else if (isInLoopPhase) {
 				setCurrentLoopSentences(shuffleArray(currentLoopSentences));
 				setCurrentSentenceIndex(0);
-				console.log("shuffled", currentLoopSentences);
 			}
 			return;
 		}
@@ -120,20 +105,10 @@ const TextAnimation = ({
 		loopSentences,
 	]);
 
-	return (
-		<div>
-			<p
-				style={{ transitionDuration: `${fadeDuration}ms` }}
-				className={
-					fadeTrue // conditional. If fade exists, then fade it out. Otherwise, don't fade it out til the step completes
-						? `transition-opacity ${fadingOut ? "opacity-0" : "opacity-100"}`
-						: ""
-				}
-			>
-				{displayedText}
-			</p>
-		</div>
-	);
+	return {
+		displayedText,
+		fadingOut,
+		fadeDuration,
+		fadeTrue,
+	};
 };
-
-export default TextAnimation;
